+++ wemake_python_styleguide/violations/complexity.py
@@ -112,7 +112,7 @@ Structure complexity
 
 """
 
-from typing_extensions import final
+from typing import final
 
 from wemake_python_styleguide.violations.base import (
     ASTViolation,