+++ wemake_python_styleguide/violations/consistency.py
@@ -158,7 +158,7 @@ Consistency checks
 
 """
 
-from typing_extensions import final
+from typing import final
 
 from wemake_python_styleguide.violations.base import (
     ASTViolation,