+++ wemake_python_styleguide/visitors/ast/blocks.py
@@ -1,8 +1,6 @@
 import ast
 from collections import defaultdict
-from typing import TypeAlias, cast
-
-from typing_extensions import final
+from typing import TypeAlias, cast, final
 
 from wemake_python_styleguide.compat.aliases import ForNodes, WithNodes
 from wemake_python_styleguide.logic import walk