+++ wemake_python_styleguide/visitors/ast/complexity/overuses.py
@@ -1,9 +1,7 @@
 import ast
 from collections import defaultdict
 from collections.abc import Callable
-from typing import ClassVar, TypeAlias
-
-from typing_extensions import final
+from typing import ClassVar, TypeAlias, final
 
 from wemake_python_styleguide.compat.aliases import FunctionNodes
 from wemake_python_styleguide.logic import source, walk