+++ wemake_python_styleguide/visitors/ast/operators.py
@@ -1,8 +1,6 @@
 import ast
 from collections.abc import Mapping
-from typing import ClassVar, TypeAlias
-
-from typing_extensions import final
+from typing import ClassVar, TypeAlias, final
 
 from wemake_python_styleguide.compat.aliases import TextNodes
 from wemake_python_styleguide.logic import walk