+++ wemake_python_styleguide/visitors/tokenize/conditions.py
@@ -1,8 +1,6 @@
 import tokenize
 from collections.abc import Sequence
-from typing import ClassVar
-
-from typing_extensions import final
+from typing import ClassVar, final
 
 from wemake_python_styleguide.violations.refactoring import (
     ImplicitElifViolation,